+++ CHANGELOG.md
@@ -78,3 +78,4 @@
 * Rename type-hint ``Schema`` to ``DocumentSchema``.
 * Add empty document class ``histore.document.mem.Schema``.
 * Change format of serialized archive JSON files.
+* Change internal representation of timestamps.